+++ src/main/java/com/company/projectoop/model/To.java
@@ -3,20 +3,26 @@ package com.company.projectoop.model;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Classe che identifica un campo della tabella
+ * @author Matteo Perozzi
+ * @author Ettore Zamponi
+ */
+
 @Component
 @Repository
 public class To {
 
-    private float to;
+    private int to;
 
     //costruttore della classe
     public To() {}
 
-    public float getTo() {
+    public int getTo() {
         return to;
     }
 
-    public void setTo(float to) {
+    public void setTo(int to) {
         this.to = to;
     }
